하네스 추가 — pytest 139개 · ruff · 커밋 전 리뷰 훅
이 레포의 사고는 거의 다 예외 없이 일어났다. WEBP를 보내면 HTTP 200이 오는데
모델은 이미지를 못 봤고, system 프롬프트 1바이트가 바뀌면 프리픽스 캐시가
죽는데 에러는 안 났고, 덜 그려진 화면을 캡처해도 아무것도 실패하지 않고
탐색만 다른 길로 샜다. "예외가 안 났다"는 아무것도 보장하지 않는다.

그래서 테스트는 전부 불변식이고, 전부 실제로 났던 사고에서 나왔다.
각 파일 상단에 어느 사고인지 적어 뒀다.

app/tests/ · 139개 · 0.8초 · 브라우저도 VLM 서버도 없이 돈다
  test_prompt.py         해시 핀 · 버전 선택 · 스키마 필드가 프롬프트에 정의됐는가
  test_imaging.py        무엇을 넣어도 JPEG · 고정 크기 · 늘리지 않고 자른다
  test_vlm.py            이미지 무시 · 캐시 미스 · 500 연속 · 잘린 출력
  test_walk.py           온 길 제외 · 막다른 길이 폴백으로 안 샘 · frontier
  test_kakao_settle.py   프레임 안정화. 1차 수정이 뚫렸던 A,A,B,B,B 시퀀스를 고정
  test_geo_providers.py  측지 왕복 · 격자 스냅 안정성
  test_config_secrets.py 키 값이 예외·진단 출력 어디에도 안 나온다

브라우저가 필요한 kakao.py 의 _settle/_await_tiles 도 커버한다. 페이지 객체의
좁은 인터페이스만 쓰므로 __init__ 을 건너뛰고 가짜 페이지를 꽂으면 Playwright
없이 프레임 시퀀스를 정확히 재현할 수 있다.

ruff (E W F I B UP SIM RUF) — 60건을 정리했다. 실질적인 것 하나:
walk() 의 cfg 기본 인자가 WalkConfig() 인스턴스 하나를 호출 간에 공유했다(B008).
E702는 껐다 — `stop_reason = "dead_end"; break` 는 종료 이유를 탈출문 옆에
붙여 두는 의도적 관용구다. bench/sweep.py 의 SYSTEM 은 바이트 고정이라
재배치 대신 noqa 를 달았다(재배치가 곧 벤치 수치 손상이다).

커밋 전 리뷰 훅 (.claude/hooks/review-on-commit.py, PreToolUse/Bash):
  1단계 ruff + pytest        실패하면 차단. 이건 판단이 아니라 사실이다
  2단계 claude -p 리뷰 에이전트  읽기 전용(plan + Read/Grep/Glob), 스테이지된 diff만
        blocking → 차단 / advisory → 통과시키되 보여준다

과잉 차단은 훅을 죽인다. block-secret-reads.py 의 첫 판이 .env 를 언급만 해도
막아서 커밋 메시지조차 못 쓰게 만들었던 교훈을 그대로 적용했다:
리뷰어가 고장나면(claude 없음·타임아웃·JSON 깨짐) 통과시키고, 차단은
blocking 에만 걸고, --no-verify 는 언제나 존중한다.

검증: 네트워크 오류를 False 로 뭉개는 코드를 일부러 스테이징했더니
"실패와 정상 부정 판정을 구분할 수 없다"로 blocking 판정. 이 레포에서
가장 비싼 버그 유형을 정확히 짚었다.

docs/12-harness.md 신규.

+++ app/trailwalk/imaging.py
@@ -55,7 +55,7 @@ def _encode(img: Image.Image) -> str:
         w, h = img.size
         want = TARGET_SIZE[0] / TARGET_SIZE[1]
         if abs(w / h - want) > 1e-3:
-            cw, ch = (int(round(h * want)), h) if w / h > want else (w, int(round(w / want)))
+            cw, ch = (round(h * want), h) if w / h > want else (w, round(w / want))
             img = img.crop(((w - cw) // 2, (h - ch) // 2, (w - cw) // 2 + cw, (h - ch) // 2 + ch))
         img = img.resize(TARGET_SIZE, Image.LANCZOS)
     buf = io.BytesIO()
